Reorganize distributed classifiers
